pyproject.toml
[project]
name = "rhaid"
version = "0.7.1"
description = "Rhaid Autofix — repo hygiene with smart AI guardrails"
readme = "README.md"
requires-python = ">=3.9"
authors = [{name="Camwood Inc."}]

dependencies = [
	"PyYAML>=6.0.1",
	"pathspec>=0.12.1",
	"rich>=13.7.1",
	"typer>=0.12.3"
]

[project.optional-dependencies]
llm = ["openai>=1.30.0"]
hcl = ["python-hcl2>=4.3.3"]
tomlkit = ["tomlkit>=0.12.0"]


[project.scripts]
rhaid = "rhaid_autofix:main"



[tool.rhaid.changelog]
"0.7.0" = [
	"Major refactor for maintainability and extensibility",
	"Dynamic plugin loading for rules/fixers",
	"VS Code extension: settings page, progress feedback",
	"Gradio app: repo URL, authentication, rule selection",
	"CI/CD: summary report artifact, improved YAML",
	"Expanded documentation and API reference"
]

[tool.setuptools.packages.find]
where = ["src"]

[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"
